refactor: split fungi_control.rs
